[Del] 필요없는 주석 삭제 (#34)
