+++ DATEROAD-iOS/DATEROAD-iOS/Presentation/CourseDetail/Cells/BringCourseCell.swift
@@ -26,15 +26,10 @@ final class BringCourseCell: BaseCollectionViewCell {
         }
     }
     
-    // MARK: - Properties
-    
-//    static let identifier: String = "BringCourseCell"
-    
     override init(frame: CGRect) {
+        
         super.init(frame: frame)
-//        setHierarchy()
-//        setLayout()
-//        setStyle()
+
         setupTapGesture()
     }
     
@@ -110,50 +105,4 @@ private extension BringCourseCell {
             likeButtonImageView.tintColor = UIColor(resource: .gray200)
         }
     }
-    
-//    override func setHierarchy() {
-//        self.addSubviews(
-//            likeButtonView,
-//            likeButtonImageView,
-//            bringCourseButton
-//        )
-//    }
-//    
-//    override func setLayout() {
-//        likeButtonView.snp.makeConstraints {
-//            $0.leading.verticalEdges.equalToSuperview()
-//            $0.width.equalTo(72)
-//        }
-//        
-//        likeButtonImageView.snp.makeConstraints {
-//            $0.center.equalTo(likeButtonView)
-//            $0.width.equalTo(20)
-//            $0.height.equalTo(18)
-//        }
-//        
-//        bringCourseButton.snp.makeConstraints {
-//            $0.trailing.verticalEdges.equalToSuperview()
-//            $0.leading.equalTo(likeButtonView.snp.trailing).offset(16)
-//        }
-//    }
-//    
-//    override func setStyle() {
-//        likeButtonView.do {
-//            $0.backgroundColor = UIColor(resource: .gray100)
-//            $0.roundCorners(cornerRadius: 14, maskedCorners: [.layerMinXMinYCorner, .layerMaxXMinYCorner, .layerMinXMaxYCorner, .layerMaxXMaxYCorner])
-//        }
-//        
-//        likeButtonImageView.do {
-//            $0.image = UIImage(resource:.heartIcon).withRenderingMode(.alwaysTemplate)
-//            $0.tintColor = UIColor(resource: .gray200)
-//        }
-//        
-//        bringCourseButton.do {
-//            $0.roundedButton(cornerRadius: 14, maskedCorners: [.layerMinXMinYCorner, .layerMaxXMinYCorner, .layerMinXMaxYCorner, .layerMaxXMaxYCorner])
-//            $0.backgroundColor = UIColor(resource: .deepPurple)
-//            $0.setTitle(StringLiterals.CourseDetail.bringCourseLabel, for: .normal)
-//            $0.setTitleColor(UIColor(resource: .drWhite), for: .normal)
-//            $0.titleLabel?.font = UIFont.suit(.body_bold_15)
-//        }
-//    }
 }
