+++ DATEROAD-iOS/DATEROAD-iOS/Presentation/CourseDetail/Cells/CoastInfoCell.swift
@@ -22,19 +22,6 @@ final class CoastInfoCell: BaseCollectionViewCell {
     // MARK: - Properties
     
     static let identifier: String = "CoastInfoCell"
-
-//    override init(frame: CGRect) {
-//        
-//        super.init(frame: frame)
-//        setHierarchy()
-//        setLayout()
-//        setStyle()
-//    }
-//    
-//    required init?(coder: NSCoder) {
-//        fatalError("init(coder:) has not been implemented")
-//    }
-    
     
     override func setHierarchy() {
         self.addSubviews(timelineBackgroundView, coastLabel)
@@ -73,41 +60,4 @@ extension CoastInfoCell {
     }
     
 }
-
-// MARK: - Private Methods
-
-private extension CoastInfoCell {
-//    
-//    func setHierarchy() {
-//        self.addSubviews(timelineBackgroundView, coastLabel)
-//    }
-//    
-//    func setLayout() {
-//        timelineBackgroundView.snp.makeConstraints {
-//            $0.leading.trailing.bottom.equalToSuperview()
-//            $0.height.equalTo(50)
-//        }
-//        
-//        coastLabel.snp.makeConstraints {
-//            $0.leading.equalToSuperview().offset(20)
-//            $0.centerY.equalTo(timelineBackgroundView)
-//        }
-//    }
-//    
-//    func setStyle() {
-//        timelineBackgroundView.do {
-//            $0.backgroundColor = UIColor(resource: .gray100)
-//            $0.layer.cornerRadius = 14
-//        }
-//        
-//        coastLabel.do {
-//            $0.text = "90,000원"
-//            $0.font = UIFont.suit(.body_bold_15)
-//            $0.textColor = UIColor(resource: .drBlack)
-//        }
-//    }
-
-}
-
-
