+++ DATEROAD-iOS/DATEROAD-iOS/Presentation/CourseDetail/Cells/MainContentsCell.swift
@@ -2,317 +2,7 @@
 //  MainContentsCell.swift
 //  DATEROAD-iOS
 //
-//  Created by 김민서 on 7/1/24.
+//  Created by 김민서 on 7/8/24.
 //
 
-import UIKit
-
-import SnapKit
-import Then
-
-
-final class MainContentsCell: BaseCollectionViewCell {
-    
-    // MARK: - UI Properties
-    
-    private let dateLabel = UILabel()
-    
-    private let titleView = UIView()
-    
-    private let titleLabel = UILabel()
-    
-    private let coastIconImageView = UIImageView(image: .coastIcon)
-    
-    private let coastLabel = UILabel()
-    
-    private let timeIconImageView = UIImageView(image: .timeIcon)
-    
-    private let timeLabel = UILabel()
-    
-    private let locationIconImageView = UIImageView(image: .locationIcon)
-    
-    private let locationLabel = UILabel()
-    
-    private let mainTextLabel = UILabel()
-    
-    
-    // MARK: - Properties
-//    
-//    static let identifier: String = "MainContentsCell"
-//    
-//    override init(frame: CGRect) {
-//        
-//        super.init(frame: frame)
-//        setHierarchy()
-//        setLayout()
-//        setStyle()
-//    }
-//    
-//    required init?(coder: NSCoder) {
-//        fatalError("init(coder:) has not been implemented")
-//    }
-    
-    override func setHierarchy() {
-        self.addSubviews(
-            dateLabel,
-            titleLabel,
-            coastLabel,
-            coastIconImageView,
-            titleView,
-            timeLabel,
-            timeIconImageView,
-            locationLabel,
-            locationIconImageView,
-            mainTextLabel
-        )
-    }
-    
-    override func setLayout() {
-        dateLabel.snp.makeConstraints {
-            $0.top.equalToSuperview().inset(23)
-            $0.leading.equalToSuperview()
-        }
-        titleView.snp.makeConstraints {
-            $0.top.equalTo(dateLabel.snp.bottom).offset(16)
-            $0.leading.trailing.equalToSuperview()
-            $0.height.equalTo(62)
-        }
-        
-        titleLabel.snp.makeConstraints {
-            $0.top.equalTo(titleView)
-            $0.leading.trailing.equalToSuperview()
-        }
-        
-        coastIconImageView.snp.makeConstraints {
-            $0.top.equalTo(titleView.snp.bottom).offset(15)
-            $0.leading.equalToSuperview()
-            $0.width.height.equalTo(14)
-        }
-        
-        coastLabel.snp.makeConstraints {
-            $0.centerY.equalTo(coastIconImageView)
-            $0.leading.equalTo(coastIconImageView.snp.trailing).offset(3.5)
-        }
-        
-        timeIconImageView.snp.makeConstraints {
-            $0.top.equalTo(coastIconImageView)
-            $0.leading.equalTo(coastLabel.snp.trailing).offset(25)
-            $0.width.height.equalTo(14)
-        }
-        
-        timeLabel.snp.makeConstraints {
-            $0.centerY.equalTo(timeIconImageView)
-            $0.leading.equalTo(timeIconImageView.snp.trailing).offset(3.5)
-        }
-        
-        locationIconImageView.snp.makeConstraints {
-            $0.top.equalTo(coastIconImageView)
-            $0.leading.equalTo(timeLabel.snp.trailing).offset(25)
-            $0.width.equalTo(12)
-            $0.height.equalTo(14)
-        }
-        
-        locationLabel.snp.makeConstraints {
-            $0.centerY.equalTo(locationIconImageView)
-            $0.leading.equalTo(locationIconImageView.snp.trailing).offset(3.5)
-        }
-        
-        mainTextLabel.snp.makeConstraints {
-            $0.top.equalTo(coastLabel.snp.bottom).offset(20)
-            $0.leading.trailing.equalToSuperview()
-            $0.bottom.equalToSuperview().inset(34)
-        }
-        
-    }
-    
-    override func setStyle() {
-        
-        dateLabel.do {
-            $0.text = "2024년 6월 27일"
-            $0.font = UIFont.suit(.body_semi_15)
-            $0.textColor = UIColor(resource: .gray400)
-        }
-        
-        titleLabel.do {
-            $0.text = "나랑 스껄 할래?"
-            $0.font = UIFont.suit(.title_extra_24)
-            $0.textColor = UIColor(resource: .drBlack)
-            $0.numberOfLines = 0
-        }
-        
-        coastLabel.do {
-            $0.text = "10만원 이하"
-            $0.font = UIFont.suit(.body_semi_15)
-            $0.textColor = UIColor(resource: .gray400)
-        }
-        
-        timeLabel.do {
-            $0.text = "6시간"
-            $0.font = UIFont.suit(.body_semi_15)
-            $0.textColor = UIColor(resource: .gray400)
-        }
-        
-        locationLabel.do {
-            $0.text = "건대/성수/왕십리"
-            $0.font = UIFont.suit(.body_semi_15)
-            $0.textColor = UIColor(resource: .gray400)
-        }
-        
-        mainTextLabel.do {
-            let paragraphStyle = NSMutableParagraphStyle()
-            paragraphStyle.lineHeightMultiple = 1.2
-            $0.attributedText = NSMutableAttributedString(string: "본문", attributes: [NSAttributedString.Key.paragraphStyle: paragraphStyle])
-
-            $0.font = UIFont.suit(.body_med_13)
-            $0.textColor = UIColor(resource: .drBlack)
-            $0.numberOfLines = 0
-       
-        }
-        
-    }
-    
-    
-}
-
-extension MainContentsCell {
-    
-    func setCell(mainContentsData: MainContents) {
-        dateLabel.text = mainContentsData.date
-        titleLabel.text = mainContentsData.title
-        coastLabel.text = mainContentsData.coast
-        timeLabel.text = mainContentsData.time
-        locationLabel.text = mainContentsData.location
-        mainTextLabel.text = mainContentsData.mainText
-    }
-}
-
-// MARK: - Private Methods
-
-private extension MainContentsCell {
-    
-//    func setHierarchy() {
-//        self.addSubviews(
-//            dateLabel,
-//            titleLabel,
-//            coastLabel,
-//            coastIconImageView,
-//            titleView,
-//            timeLabel,
-//            timeIconImageView,
-//            locationLabel,
-//            locationIconImageView,
-//            mainTextLabel
-//        )
-//    }
-//    
-//    func setLayout() {
-//        dateLabel.snp.makeConstraints {
-//            $0.top.equalToSuperview().inset(23)
-//            $0.leading.equalToSuperview()
-//        }
-//        titleView.snp.makeConstraints {
-//            $0.top.equalTo(dateLabel.snp.bottom).offset(16)
-//            $0.leading.trailing.equalToSuperview()
-//            $0.height.equalTo(62)
-//        }
-//        
-//        titleLabel.snp.makeConstraints {
-//            $0.top.equalTo(titleView)
-//            $0.leading.trailing.equalToSuperview()
-//        }
-//        
-//        coastIconImageView.snp.makeConstraints {
-//            $0.top.equalTo(titleView.snp.bottom).offset(15)
-//            $0.leading.equalToSuperview()
-//            $0.width.height.equalTo(14)
-//        }
-//        
-//        coastLabel.snp.makeConstraints {
-//            $0.centerY.equalTo(coastIconImageView)
-//            $0.leading.equalTo(coastIconImageView.snp.trailing).offset(3.5)
-//        }
-//        
-//        timeIconImageView.snp.makeConstraints {
-//            $0.top.equalTo(coastIconImageView)
-//            $0.leading.equalTo(coastLabel.snp.trailing).offset(25)
-//            $0.width.height.equalTo(14)
-//        }
-//        
-//        timeLabel.snp.makeConstraints {
-//            $0.centerY.equalTo(timeIconImageView)
-//            $0.leading.equalTo(timeIconImageView.snp.trailing).offset(3.5)
-//        }
-//        
-//        locationIconImageView.snp.makeConstraints {
-//            $0.top.equalTo(coastIconImageView)
-//            $0.leading.equalTo(timeLabel.snp.trailing).offset(25)
-//            $0.width.equalTo(12)
-//            $0.height.equalTo(14)
-//        }
-//        
-//        locationLabel.snp.makeConstraints {
-//            $0.centerY.equalTo(locationIconImageView)
-//            $0.leading.equalTo(locationIconImageView.snp.trailing).offset(3.5)
-//        }
-//        
-//        mainTextLabel.snp.makeConstraints {
-//            $0.top.equalTo(coastLabel.snp.bottom).offset(20)
-//            $0.leading.trailing.equalToSuperview()
-//            $0.bottom.equalToSuperview().inset(34)
-//        }
-//        
-//    }
-//    
-//    func setStyle() {
-//        
-//        dateLabel.do {
-//            $0.text = "2024년 6월 27일"
-//            $0.font = UIFont.suit(.body_semi_15)
-//            $0.textColor = UIColor(resource: .gray400)
-//        }
-//        
-//        titleLabel.do {
-//            $0.text = "나랑 스껄 할래?"
-//            $0.font = UIFont.suit(.title_extra_24)
-//            $0.textColor = UIColor(resource: .drBlack)
-//            $0.numberOfLines = 0
-//        }
-//        
-//        coastLabel.do {
-//            $0.text = "10만원 이하"
-//            $0.font = UIFont.suit(.body_semi_15)
-//            $0.textColor = UIColor(resource: .gray400)
-//        }
-//        
-//        timeLabel.do {
-//            $0.text = "6시간"
-//            $0.font = UIFont.suit(.body_semi_15)
-//            $0.textColor = UIColor(resource: .gray400)
-//        }
-//        
-//        locationLabel.do {
-//            $0.text = "건대/성수/왕십리"
-//            $0.font = UIFont.suit(.body_semi_15)
-//            $0.textColor = UIColor(resource: .gray400)
-//        }
-//        
-//        mainTextLabel.do {
-//            let paragraphStyle = NSMutableParagraphStyle()
-//            paragraphStyle.lineHeightMultiple = 1.2
-//            $0.attributedText = NSMutableAttributedString(string: "본문", attributes: [NSAttributedString.Key.paragraphStyle: paragraphStyle])
-//
-//            $0.font = UIFont.suit(.body_med_13)
-//            $0.textColor = UIColor(resource: .drBlack)
-//            $0.numberOfLines = 0
-//       
-//        }
-//        
-//    }
-//    
-//    
-    
-}
-
-
-
-
+import Foundation
